Got cross db queries working between sql and mongo
Had to compromise on my idealized view of every module using a factory pattern and taking its dependencies as arguments. This was because of a circular reference issue that emerged in the sql GraphQL type definitions.

For instance, The two review types need to reference Reviewer, but Reviewer needs to to reference Reviews, which needs to reference the two review types.

This recursive relationship is fine if the modules return static objects.

The factory pattern, however, returns functions that return static objects, and those functions need other type definitions to be passed as arguments, which can't exist yet because they need to created using type def.... I'm intentionally trailing off there because it is an unsolved issue for me currently.

As a result this small set of files imports their GraphQL standard types from the global namespace (bleh)